requirements.txt
pyspark==3.5.0
pymysql==1.1.0
mysql-connector-python==8.2.0
pandas==2.1.4
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3
jsonschema==4.20.0
loguru==0.7.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-spark==0.6.0
numpy==1.26.2
scipy==1.11.4
requests==2.31.0
tqdm==4.66.1
black==23.12.1
flake8==7.0.0
